Adds ubiquity for hard disk installation
Adds ubiquity to be installed via /scripts/chroot.sh in the chroot env
so that it will be present on the live cd. Modifies loopback.cfg of
grub so that `Install Meilix` becomes visible which runs live
installation along `--only-ubiquity` parameter which fires up ubiquity.

Adds slides which will be displayed by ubiquity as well the text displayed
during the installation procedure to the `/ubiquity-slidehsow/` folder
which overrides the files in `/usr/share/ubiquity-slideshow/` to display
slides specific to meilix. This change is reflected in build.sh

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>